+++ Runtime/UI/AButtonController.cs
@@ -92,7 +92,7 @@ namespace NineHundredLbs.Controly.UI
         /// Toggles the interactability of <see cref="UIButton"/> to the given <paramref name="value"/>.
         /// </summary>
         /// <param name="value">Whether to toggle to interactable (true) or uninteractable (false).</param>
-        public void ToggleInteractability(bool value)
+        public virtual void ToggleInteractability(bool value)
         {
             UIButton.Interactable = value;
         }
@@ -222,7 +222,7 @@ namespace NineHundredLbs.Controly.UI
         /// Toggles the interactability of <see cref="UIButton"/> to the given <paramref name="value"/>.
         /// </summary>
         /// <param name="value">Whether to toggle to interactable (true) or uninteractable (false).</param>
-        public void ToggleInteractability(bool value)
+        public virtual void ToggleInteractability(bool value)
         {
             UIButton.Interactable = value;
         }